# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "KAVACH \u2014 AI-Powered DevSecOps Platform"
  description: "KAVACH scans banking application repositories for vulnerabilities,\
    \ calculates a Banking Risk Score (BRS), maps findings to regulatory controls\
    \ (RBI / PCI-DSS / SWIFT), and generates AI-powered audit reports.\n\n## Pipeline\n\
    `GitHub webhook` \u2192 `Create Scan Job` \u2192 `Queue` \u2192 `Distributed workers`\
    \ \u2192 `Aggregation` \u2192 `Risk Engine` \u2192 `Compliance Engine` \u2192\
    \ `AI Explanation` \u2192 `Reports` \u2192 `Dashboard` \u2192 `Notifications`\
    \ \u2192 `Archive`\n\nA scan enters this pipeline either via `POST /scan/repository`\
    \ / `POST /scan` (an authenticated user submitting a URL or archive) or `POST\
    \ /webhooks/github` (an unauthenticated, HMAC-verified push event) \u2014 both\
    \ converge on the same queue \u2192 distributed-worker \u2192 aggregation path,\
    \ so everything downstream behaves identically regardless of how the scan was\
    \ triggered."
  contact:
    name: KAVACH Platform Team
    url: https://github.com/kavach-platform/kavach
  license:
    name: "Proprietary \u2014 All Rights Reserved"
  version: 1.0.0
paths:
...
  /api/v1/analytics/my-activity:
    get:
      tags:
      - Analytics
      summary: Get My Activity
      description: "A user's own scan activity \u2014 always their own data, so any\n\
        authenticated active user can read it (no extra permission beyond\nbeing logged\
        \ in; there's nothing to over-privilege here)."
      operationId: get_my_activity_api_v1_analytics_my_activity_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyActivitySummary'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/analytics/team-activity:
    get:
      tags:
      - Analytics
      summary: Get Team Activity
      description: "Org-wide scan activity broken down per user \u2014 Security Manager/Admin\
        \ only."
      operationId: get_team_activity_api_v1_analytics_team_activity_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamActivitySummary'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/assistant/chat:
    post:
      tags:
      - AI Assistant
      summary: Chat
      description: "Server-Sent Events stream. Event sequence:\n  \"retrieval\"  \
        \         \u2014 always sent first: retrieved_count, confidence,\n       \
        \                    and the full citation list (document, page,\n       \
        \                    section, similarity score, excerpt).\n  \"insufficient_context\"\
        \ \u2014 sent INSTEAD of any \"token\" events when the\n                 \
        \          confidence gate fails; the LLM is never\n                     \
        \      called in this case.\n  \"token\"                \u2014 one or more,\
        \ the streamed answer text.\n  \"done\"                 \u2014 always sent\
        \ last (after \"token\"s, never after\n                           \"insufficient_context\"\
        ): confidence,\n                           retrieved_count, latency_ms.\n\
        \  \"error\"                 \u2014 only if the stream fails part-way through."
      operationId: chat_api_v1_assistant_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/executive-intelligence/ask:
    post:
      tags:
      - Executive Intelligence
      summary: Ask Executive Intelligence
      description: "SSE stream. Event sequence:\n  \"evidence\"             \u2014\
        \ always first: the deterministic scan-history\n                         \
        \  snapshot (see evidence_service.py) + any\n                           supplementary\
        \ knowledge-base citations.\n  \"insufficient_context\" \u2014 sent INSTEAD\
        \ of any \"token\"s when there is\n                           neither scan\
        \ history nor KB support at all\n                           (an empty, brand-new\
        \ deployment) \u2014 the LLM\n                           is never called in\
        \ this case.\n  \"token\"                 \u2014 one or more, the streamed\
        \ answer text.\n  \"done\"                  \u2014 latency, always last after\
        \ \"token\"s.\n  \"error\"                  \u2014 only if the stream fails\
        \ part-way through."
      operationId: ask_executive_intelligence_api_v1_executive_intelligence_ask_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutiveAskRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/executive-intelligence/export-pdf:
    post:
      tags:
      - Executive Intelligence
      summary: Export Executive Intelligence Pdf
      description: "Renders exactly the question/answer/evidence/citations passed\
        \ in \u2014 see\napp/services/executive_intelligence/pdf_export.py's docstring\
        \ for why\nthis never recomputes anything server-side at export time."
      operationId: export_executive_intelligence_pdf_api_v1_executive_intelligence_export_pdf_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutivePdfExportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    ChatMessage:
      properties:
        role:
          type: string
          enum:
          - user
          - assistant
          title: Role
        content:
          type: string
          maxLength: 8000
          title: Content
      type: object
      required:
      - role
      - content
      title: ChatMessage
    ChatRequest:
      properties:
        message:
          type: string
          maxLength: 4000
          minLength: 1
          title: Message
        history:
          items:
            $ref: '#/components/schemas/ChatMessage'
          type: array
          maxItems: 50
          title: History
      type: object
      required:
      - message
      title: ChatRequest
...
    ComplianceFrameworkEvidenceSchema:
      properties:
        framework_key:
          type: string
          title: Framework Key
        framework_name:
          type: string
          title: Framework Name
        compliant_repo_count:
          type: integer
          title: Compliant Repo Count
        non_compliant_repo_count:
          type: integer
          title: Non Compliant Repo Count
        total_violations:
          type: integer
          title: Total Violations
      type: object
      required:
      - framework_key
      - framework_name
      - compliant_repo_count
      - non_compliant_repo_count
      - total_violations
      title: ComplianceFrameworkEvidenceSchema
...
    EvidenceSnapshotSchema:
      properties:
        generated_at:
          type: string
          title: Generated At
        total_repositories:
          type: integer
          title: Total Repositories
        total_completed_scans:
          type: integer
          title: Total Completed Scans
        total_findings:
          type: integer
          title: Total Findings
        findings_by_severity:
          additionalProperties:
            type: integer
          type: object
          title: Findings By Severity
        portfolio_average_brs:
          anyOf:
          - type: number
          - type: 'null'
          title: Portfolio Average Brs
        top_risk_repositories:
          items:
            $ref: '#/components/schemas/RepositoryRiskEvidenceSchema'
          type: array
          title: Top Risk Repositories
        compliance_by_framework:
          items:
            $ref: '#/components/schemas/ComplianceFrameworkEvidenceSchema'
          type: array
          title: Compliance By Framework
        weekly_trend:
          items:
            $ref: '#/components/schemas/WeeklyTrendPointSchema'
          type: array
          title: Weekly Trend
        week_over_week:
          anyOf:
          - $ref: '#/components/schemas/WeekOverWeekDeltaSchema'
          - type: 'null'
      type: object
      required:
      - generated_at
      - total_repositories
      - total_completed_scans
      - total_findings
      title: EvidenceSnapshotSchema
    ExecutiveAskRequest:
      properties:
        question:
          type: string
          maxLength: 2000
          minLength: 1
          title: Question
        history:
          items:
            $ref: '#/components/schemas/ChatMessage'
          type: array
          maxItems: 50
          title: History
      type: object
      required:
      - question
      title: ExecutiveAskRequest
    ExecutiveCitation:
      properties:
        document_id:
          type: string
          title: Document Id
        filename:
          type: string
          title: Filename
        page_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Page Number
        section_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Section Path
        heading:
          anyOf:
          - type: string
          - type: 'null'
          title: Heading
        similarity_score:
          type: number
          title: Similarity Score
        excerpt:
          type: string
          maxLength: 4000
          title: Excerpt
      type: object
      required:
      - document_id
      - filename
      - similarity_score
      - excerpt
      title: ExecutiveCitation
    ExecutivePdfExportRequest:
      properties:
        question:
          type: string
          maxLength: 2000
          title: Question
        answer:
          type: string
          maxLength: 20000
          title: Answer
        evidence:
          $ref: '#/components/schemas/EvidenceSnapshotSchema'
        citations:
          items:
            $ref: '#/components/schemas/ExecutiveCitation'
          type: array
          maxItems: 20
          title: Citations
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
      type: object
      required:
      - question
      - answer
      - evidence
      title: ExecutivePdfExportRequest
      description: "Carries back exactly what the client already displayed \u2014\
        \ see\napp/services/executive_intelligence/pdf_export.py's docstring for why\n\
        this is never recomputed server-side at export time."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MyActivitySummary:
      properties:
        total_scans:
          type: integer
          title: Total Scans
        scans_by_status:
          additionalProperties:
            type: integer
          type: object
          title: Scans By Status
        total_findings:
          type: integer
          title: Total Findings
        findings_by_severity:
          additionalProperties:
            type: integer
          type: object
          title: Findings By Severity
        average_scan_duration_seconds:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Scan Duration Seconds
        average_brs_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Brs Score
        recent_scans:
          items:
            $ref: '#/components/schemas/RecentScanSummary'
          type: array
          title: Recent Scans
      type: object
      required:
      - total_scans
      - scans_by_status
      - total_findings
      - findings_by_severity
      - recent_scans
      title: MyActivitySummary
      description: "A Security Analyst's own workload/efficiency view \u2014 see\n\
        ROLE_DISPLAY_NAMES in app/auth/permissions.py for the role mapping."
    RecentScanSummary:
      properties:
        scan_job_id:
          type: string
          format: uuid
          title: Scan Job Id
        repository_name:
          type: string
          title: Repository Name
        status:
          type: string
          title: Status
        brs_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Brs Score
        brs_risk_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Brs Risk Level
        finished_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Finished At
      type: object
      required:
      - scan_job_id
      - repository_name
      - status
      title: RecentScanSummary
...
    RepositoryRiskEvidenceSchema:
      properties:
        repository_id:
          type: string
          title: Repository Id
        repository_name:
          type: string
          title: Repository Name
        latest_brs_score:
          type: number
          title: Latest Brs Score
        latest_brs_risk_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Latest Brs Risk Level
        latest_scan_finished_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Latest Scan Finished At
      type: object
      required:
      - repository_id
      - repository_name
      - latest_brs_score
      title: RepositoryRiskEvidenceSchema
...
    TeamActivitySummary:
      properties:
        total_scans:
          type: integer
          title: Total Scans
        total_findings:
          type: integer
          title: Total Findings
        members:
          items:
            $ref: '#/components/schemas/TeamMemberActivity'
          type: array
          title: Members
      type: object
      required:
      - total_scans
      - total_findings
      - members
      title: TeamActivitySummary
      description: "Org-wide scan-activity aggregation for a Security Manager \u2014\
        \ how much\ngot scanned, by whom, what it found. Deliberately not the same\
        \ thing as\nthe audit log (who changed what) \u2014 see Permission.TEAM_ANALYTICS_READ's\n\
        docstring."
    TeamMemberActivity:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        email:
          type: string
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        total_scans:
          type: integer
          title: Total Scans
        total_findings:
          type: integer
          title: Total Findings
        average_brs_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Brs Score
      type: object
      required:
      - user_id
      - email
      - total_scans
      - total_findings
      title: TeamMemberActivity
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WeekOverWeekDeltaSchema:
      properties:
        scans_this_week:
          type: integer
          title: Scans This Week
        scans_last_week:
          type: integer
          title: Scans Last Week
        findings_this_week:
          type: integer
          title: Findings This Week
        findings_last_week:
          type: integer
          title: Findings Last Week
        average_brs_this_week:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Brs This Week
        average_brs_last_week:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Brs Last Week
      type: object
      required:
      - scans_this_week
      - scans_last_week
      - findings_this_week
      - findings_last_week
      title: WeekOverWeekDeltaSchema
    WeeklyTrendPointSchema:
      properties:
        week_start:
          type: string
          title: Week Start
        scan_count:
          type: integer
          title: Scan Count
        average_brs:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Brs
        critical_high_findings:
          type: integer
          title: Critical High Findings
      type: object
      required:
      - week_start
      - scan_count
      - critical_high_findings
      title: WeeklyTrendPointSchema
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...